Preserve the commit list state in case a candidate is rejected.
Prior to this change, a candidate was selected by finding the repo that
would result in the least movement in the configuration. The chosen repo
would have its commit list advanced, along with those of its dependencies
and then the new state of the configuration would be checked. If the
configuration was rejected, the state of the commit lists wasn't restored,
so it was eliminating potential candidates that could be formed by
advancing other repos while keeping the chosen repo at its current commit.

This was resulting in rolls into some repos frequently getting stuck.

With this change, a temporary copy of the state is modified for testing
out a candidate. Only when a candidate is accepted is the original state
updated. Additionally, instead of finding a single candidate for a given
set of revisions, all of the possible candidates are returned in a list
sorted in the order that the candidates should be tested. This is
necessary to avoid repeatedly testing the same failed candidate.

Bug: 1171386
Change-Id: Iaefec97f83bd4fac595bfca6f43fc177cbc66d5d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2653978
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>