[file] Catch IOErrors
In Python <3.3, trying to access a non-existent file raises an IOError, not
an OSError. IOError and OSError are the two subclasses of
EnvironmentError, and they both have the `errno` attribute, so we can
catch `EnvironmentError` here to make sure we catch an exception for
nonexistent files.

This can be changed back to OSError in Python 3, since IOError was
merged into OSError in Python 3.3 and EnvironmentError became an alias
for OSError.

Change-Id: I233652f68f4d095c9e52ae7b2f426b5664250d2c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2937257
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>