[file] Handle EnvironmentError with errno=None
I incorrectly assumed in crrev.com/c/2937257 that errno is always set
for EnvironmentErrors. However, it appears that is not true in Python 2;
for example, if you try call `os.rename()` for a regular file when there
already exists a directory with the new name, it will raise an IOError
without an errno. So we need to handle that case and not try to look up
the errno name for None, which causes KeyErrors.

Fortunately, this bug basically only manifests where there's a bug in
the calling recipe and the code would be likely to fail anyway. It just
makes the error message harder to understand, e.g.
https://ci.chromium.org/b/8845400030905776448.

Change-Id: I5eaff2906e5ee7ac45cfccb2d678d88c7ece9043
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/2940925
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>