Add function to compute effective python compatibility.
As python version compatibility improves, this will allow us to see which
recipes can (nominally) run on python3, and will also indicate where we
should not remove py2 compatibility yet.

This will also be useful for `recipes.py test` to determine if python3
tests should be expected to pass or not yet (e.g. if the claimed
compatibility for a given recipe is 'PY2+3', but depends on a 'PY2'
module, `recipes.py test` shouldn't expect that the 'PY3' tests should
pass).

R=yiwzhang, yuanjunh

Bug: 1211650
Change-Id: I31cdc767ab067986815ba83bad1277b82e35536c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3002697
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>