[proto] Respect codec in logs
It's confusing to present a textpb log for a proto that's read from or
written to a jsonpb file. So instead, use the same codec as the file
contents for the log (unless the codec is binary, in which case we'll
use textproto for readability).

Also change the default jsonpb indentation to 2 spaces for readability.

Recipe-Nontrivial-Roll: build
Recipe-Nontrivial-Roll: chromiumos
Recipe-Nontrivial-Roll: fuchsia
Change-Id: I0c49aa3a2b98bdaf3daa5e6fea3ee10be4557d9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3135722
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>