[swarming] Avoid comparing string to None
A TaskResult corresponding to a task that is incomplete, or that
couldn't be fetched, will have `None` as its name:
https://chromium.googlesource.com/infra/luci/recipes-py/+/6e3762eb934f77d76f11b1da6c86ec1f0f1a9407/recipe_modules/swarming/api.py#1059

`None` is no longer comparable with strings in Python 3, so it's invalid
to sort a list that contains both None and strings.

The logic for handling different task states is really ugly and should
probably be cleaned up at some point, but in the meantime this is the
fastest way to fix the python version incompatibility issue.

Bug: 1239059
Change-Id: I5a4d31ddd63bea6cfae008240c059e94008b4171
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3141777
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>