Enable overriding the status in step.raise_on_failure.
Existing code using active_result sometimes changes the status of a step
in a finally block, which means that the exception raised is dependent
on the initial status of the step result rather than the status at the
time the exception would be raised. This adds support for an additional
argument to raise_on_failure that allows for specifying the status that
the exception will be based on. The caller can save the initial value of
the status and pass it to raise_on_failure in order to have the
exception type reflect that status.

Bug: 1246180
Change-Id: I8890354edccc6cabb5fb4fabb11a3727f949e294
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3149847
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>