When advancing repos, advance to the next roll candidate commit.
Prior to this change, when a repo was advanced, it was advanced by a
single commit. This meant that if there were non-roll candidate commits,
a candidate config would be created for them and then additional
iterations would be required to advance the repo to the candidate
commit. Since these commits are not of interest for rolling, spending
time computing and checking these candidate configs provides limited
value.

Change-Id: Iaa23e0515a8f983b0f04ec5afc02684b1a80135e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3155000
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>