Reland "[recipe-py] expose resultdb info in LUCI_CONTEXT"
This is a reland of crrev.com/c/3153276 with a fix: update token should not be hidden in api.context.luci_context.

In the previous CL I hid resultdb update_token because I thought it will be revealed in execution details.

But it turned out that the LUCI_CONTEXT file will be updated by api.context.luci_context[1] which means update_token will be lost. Also [2] controls what sections of LUCI_CONTEXT to show in execution details so update token will not be leaked.

[1] https://source.chromium.org/chromium/_/chromium/infra/luci/recipes-py/+/17c88f22dfef427f9e085e2e7078cbeaaca5be6c:recipe_engine/internal/engine.py;l=810;drc=ffe17790c212701a06373ab9fd0ce3a54dab9173
[2] https://source.chromium.org/chromium/_/chromium/infra/luci/recipes-py/+/17c88f22dfef427f9e085e2e7078cbeaaca5be6c:recipe_engine/internal/engine.py;l=1024;drc=ffe17790c212701a06373ab9fd0ce3a54dab9173

Bug: 1245438
Change-Id: I07aa291594699a6bb6105ac6ce5b29b93782144a
Recipe-Manual-Change: fuchsia
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3162665
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>