[swarming] Infer server from env
This way a builder will no longer need to set its Swarming server with
the "$recipe_engine/swarming" property. If the property is unset, the
a build will default to launching tasks on the same Swarming server that
the build is running on.

Bug: 1148539
Change-Id: If72b10bef7b2dabb1b21482ea3752326bb2fce24
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3173593
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>