[json] Respect sort_keys kwarg
There are times when we want to pass in an OrderedDict to
`api.json.dumps` with the keys in a custom order, but setting
`sort_keys` to always be True means it's impossible to dump a dict while
preserving the original order.

Use `setdefault` instead so if someone sets `sort_keys=False` it will be
respected.

Change-Id: I59dfda2d9a1df4c01cd4feb8680cba523608e18e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3217551
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>