[cipd] Relax list type check
If `refs` or a similar field comes from an input property in a recipe
that uses proto properties, its type will be a proto-specific
RepeatedScalarFieldContainer rather than a normal list or tuple.
RepeatedScalarFieldContainer inherits from collections.abc.Sequence (as
do list and tuple) so we can check that the type is Sequence instead,
though we must explicitly filter out strings, since we don't want to
treat strings as sequences here.

As long as I'm here, also check for Mapping instead of dict for fields
that are expected to be dicts.

Change-Id: I347efb95f1b7daf59c6db49f58387068a47d33ea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3245910
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>