Address behavioral differences of random between python versions.
The random module's implementation is changed between python 2.7 and
python 3.8, which makes the current implementation of the random recipe
module unsuitable for expectation tests where the expectation
incorporates random data. In python 3.8, the randomness when generating
integers comes from getrandbits() instead of random(), which results in
different values.

This change defines a custom subclass of random.Random in python3 that
will cause random() to be used as the source of randomness to match
python2.

Change-Id: Ib26b2d6bbc1b27bedc3a8fa0255821a69eda0a04
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3243938
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>