Don't enqueue the py3 version of a test until the py2 version completes.
This change makes it so that for tests that are run against both
versions, the python3 version won't be enqueued until after the python2
version completes. This ensures that if the expectation file gets
updated by the python2 version, then the python3 version will see the
updated contents. This in turn prevents the need to run train twice to
get a passing run when expectation files change.

Bug: 1147793
Change-Id: Iadf1f336aa55690022075c854488fc9681b4612e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3270572
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>