[buildbucket] Allow setting empty gerrit_changes
Previously it was impossible to set gerrit_changes to an empty list
because the module overrode the argument if it was any falsy value. But
it's possible for a try builder to be triggered manually and not have
any gerrit_changes attached, so its a case were care about testing.

Change-Id: I9d516030512d11e42937afdc025e6bdd955e8fb4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3271234
Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>