[futures] Add BoundedSemaphore and example.
This can be used safely to synchronize certain sections of the recipe
within different greenlets.

The impetus is similar to the included example where a bunch of greenlets
need to run a "critical section" of sorts, with the rest of the greenlet
running independently.

R=gbeaty, yiwzhang, yuanjunh

Change-Id: I653efe8fc9864d0584c53cfb6e61c0d601300b70
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3315690
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>