Prevent spurious GitFetchErrors in CommitList.from_backend.
Currently, the recipe roller occasionally fails when creating the commit
lists with an error indicating an invalid range when the range is in
fact valid. The reason for this is that when the commit lists are
created, the repositories have already been fetched from the remote
repository but then the revisions are resolved against the remote
repository and no further fetching happens. This means that if the
remote ref has changed since being fetched, the resolved refspec will be
for a commit not contained in the local repository.

This change modifies the Backend.updates method to take a refspec and a
starting revision instead of two revisions. The GitBackend
implementation of updates will resolve the refspec and if the revision
isn't contained in local repo, it will fetch the refspec.

Bug: 1313964
Change-Id: I5b388cd25c560700a969a325b846a5390cffdf02
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3582601
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>