Add timing information option to the test command
It can be helpful to see how long certain tests take to execute. This
CL adds an argument to the testing subcommand which outputs timing
information for each test to a file. This is pretty barebones, but
was enough for me to discover and fix an issue in
https://crrev.com/c/3597036.

This also fixes a small typo in a function name which I noticed while
reading through this code.

Change-Id: I6f584a82996c10ec1eedd0c0d1e23dc48009c968
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3597673
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>