[snoopy] Create a luci's bcid recipe_module.
This recipe module will be used to interact with local luci provenance
server to self report critical events during a task. Previously this was
in infra repo under a different name, but for making it universally
available, we are adding it here too.

Real change in this CL is none, the older version will be removed once
all users are moved off of that.

R=jclinton@google.com

Bug:1269830
Change-Id: I12cc073cd0bc28bb6e15d6d9a4aa5391235409e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3614860
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>