Inhibit displaying coverage information on python3 failures with --stop.
Currently if the --stop flag is passed and a failure occurs on a python2
tests, the coverage information is not displayed because the
execute_queue function won't save the coverage data in the event of a
failure, so the reporter won't report anything. However, if a failure
occurs in a python3 test, it ends up displaying incomplete coverage
information because the coverage data for python2 tests will be saved.
This changes it to track if there were failures for either python
version and prevents the coverage report from being display if there
were any failures when the --stop flag is set.

Change-Id: Iac4bb1b75091f8e7a7ba89ac2d0a2331b4edc7d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3645569
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>