[lint] Fix Py3 'Disallowed import' failures
The result of the map() call is a generator in Python 3, so it was being
consumed by the loops in ImportsTest(), leading to many additional
'Disallowed import' failures. Explicitly convert the result to a tuple
so this doesn't happen anymore.

Also a minor cleanup in ImportsTest().

Change-Id: Ic84f9532d83812b1866bb3974e7f91689d619e31
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3785244
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Rob Mohr <mohrr@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>