resultdb: Ensure Artifacts.contents is binary
When the resultdb module is used by recipes running Python 3, calling
`get()` on the contents field will return a string value, rather than a
binary value. This change ensures that the field is coerced to the
correct type before trying to instantiate the proto.

TEST=./recipes.py test train

Change-Id: I61b320a2d760d8ba638a6afbd2fe128bcecfc578
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3791616
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Chris Gerber <gerb@google.com>