Don't print None-valued rendered steps.
The _print_step function asserts that the step is a Step object, so it
raises an exception if it is passed a None. Previously, _print_step was
only called when there were no render errors since render errors and
rendered steps were mutually exclusive. https://crrev.com/c/3807022
changed it so that a rendered step was returned some times when there
was a render error and _print_step was being called unconditionally, so
if the cwd of the command is wrong, it will call _print_step with a None
for the rendered step. This changes it so _print_step is only called for
non-None rendered steps. It also adds a message to the exception to
provide more debugging information if necessary in the future.

Change-Id: I0c4160fedda40a5c543a8bd90a13f8f22256f1b7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3887560
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>