switch to inspect.getfullargspec in py3
`inspect.getargspec` is deprecated and should be replaced with
`inspect.getfullargspec`. Unfortunately, that is not available in
py2 so this check is gated by system version. This was discovered
while adding typehints to recipes (the exact issue is adding a
typehint to `RunSteps` will throw an error `Function has keyword-
only parameters or annotations.`).

Change-Id: I9d34860b0406ce5e63687b3b1e12625f6eeff237
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4013415
Commit-Queue: Benjamin Shai <bshai@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>