[unittests] Remove hack to skip proto compilation
https://crrev.com/c/1626066 added a hacky optimization to
fake_recipe_deps.py to skip compiling protos for fake recipe repos that
don't contain any proto files. However, this means that the .recipe_deps
directory in the fake repo is not actually valid. So for the tests to
pass, the `bundle` command had to use an ugly hack to get the proto path
regardless of whether the fake repo had compiled its protos or not.

The optimization doesn't seem worth the complexity of the hack,
especially when the associated code is due to become even more
complicated as a result of the linked bugs, so it should be removed.

This may cause a small slowdown in tests that use fake_recipe_deps,
which will now always need to compile the recipe engine protos.

Bug: 1310557, fuchsia:103473
Change-Id: Iec45767871178bcf98baddce0a57b5f859a9e4ce
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4338080
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Oliver Newman <olivernewman@google.com>