Make status like other TestData, have expect_exception imply status.
This makes the status kwarg behave like a default expected_status value,
rather than as a non-add-able value (which will definitely help some
chrome browser tests be a bit nicer to update).

It also makes expect_exception include a default status of
INFRA_FAILURE, which brings the number of out-of-spec tests in the
recipe engine repo down to only ~50 (and, I expect, will be similar in
other repos which used this feature before, too).

Took the opportunity to brush up the doc strings a bit, as well (such as
they are).

Finally, this changes check_status to use the real RawResult object
with the real status field, rather than the approximated value in
$result which is derived from this.

R=bshai, gbeaty

Bug: 1010715
Recipe-Manual-Change: fuchsia
Change-Id: Ic24fe73f2c647fb57e2a5cf6703ba710179ceacc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4375418
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Benjamin Shai <bshai@google.com>