[StepPresentation] Give logs with bytes a well defined rendering.
This will use `backslashencode` to replace invalid UTF-8 byte sequences
with the equivalent `\uXXXX` sequence. This means that if the recipe
author adds binary data (bytes) to a log, it will actually render to
UTF-8 instead of raising an exception deep in the underbelly of the
engine when it attempts to write the data to a Logdog stream.

Also add a bit of documentation to StepPresentation.

R=bryner, bshai, yiwzhang

Recipe-Nontrivial-Roll: build
Change-Id: I16d815c3560773dad6ed3926b303eee4b49f249b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4406708
Reviewed-by: Benjamin Shai <bshai@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>