[path] Only allow dynamic paths to be set once.
The checkout dynamic path is effectively a global variable right
now, which leads to complex implementation details in the engine.

Making dynamic paths (of which "checkout" seems to be the ONLY one
now, yay!) write-once will allow a lot of cleanup work to happen.

If you find this CL because of an error that's popping up now,
you'll need to remove any duplicate assignments to
`api.path["checkout"]` in your recipe. If you are using the
`bot_update` module from depot_tools, this can be done by assigning
to `api.path["checkout"]` prior to calling `bot_update`.

Adds a comment to remove all 'config' related stuff from path; it
appears that there is exactly one config, globally, for the path
module, so this is a perfect time to start ripping it out.

R=fancl, gavinmak, gbeaty, yiwzhang

Bug: 1457059
Change-Id: I14771325318abe5235417e1fde38b15e280757e3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4637773
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Gavin Mak <gavinmak@google.com>