[dev] Rework python3 symlink to full multi-virtualenv support.
The previous iteration (of a _dev folder with a python3 symlink) worked
for some IDEs, but didn't work with e.g. pyright configuration options.
It also had the problem that the symlink would churn when using different RECIPE_DEBUGGER environment variable values.

This CL reworks the integration to generate a _venv subdirectory with a
set of stable symlinks to the actual virtualenv roots. This is compatible with existing IDEs (e.g.
/path/to/.recipe_deps/_venv/normal/bin/python3) as well as pyright.

Because pathlib on python3.8 supports symlinks on Windows (with some
caveats), this CL also lifts the 'not on Windows' restriction, though
I have not tested this on Windows to see if it works as expected.

I've removed the internal `skip_dev` flag because I'm contemplating a
totally different way to expose type stubs, etc. which should unify the
existing _pb3 import path with one that will work with RECIPE_MODULES
and RECIPES imports as well.

R=bshai, fancl

Change-Id: I3b177e5f5ca10b8ac3b9904eb485bf8647eb2146
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4655835
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Benjamin Shai <bshai@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>