[proto] Modernize proto generation.
This cleans up the proto generation code to be a bit more robust,
removes unnecessary __init__.py files, and also generate .pyi type
stubs.

This is the first change on the way to exporting type information in
a stable way for IDEs and other python editing environments, however it
is not a complete solution, which is why this CL doesn't change the
documentation.

R=bshai, fancl

Change-Id: I364432d20ef0d5d2c63c8e0b30a18e4bdd235a41
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/4651464
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>