Enable setting a retcode for a canceled step.
Steps that are canceled will generally result in a completed process
with a retcode, so this enables setting it when creating test data for a
step and having it propagated to the ExecutionResult for the test.
Because there are potentially situations where killing a canceled step
does not work and there is no retcode, the default behavior of None will
be preserved in that case. In order to allow for a None value to be
returned from retcode, the StepData.retcode property has been updated to
not provide a 0 default value and instead the sim runner will do that
for steps that are complete regularly.

Change-Id: I6d942e37de8da9e94f1c93d94407b2973efb2f98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5051061
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>