[path] Completely remove "dynamic_paths" config.
This is the next step to removing the 'config' from the path module.

The dynamic_paths config dict was originally added as a way to have
paths which could be set at runtime in the recipes. Unfortunately,
this is basically a glorified global variable and a bad idea.
Thankfully, this doesn't seem like it is currently used for anything
other than 'checkout'.

This CL gives checkout an explicit API, and removes dynamic_paths.

Subsequent CLs will add similar explicit (read only) APIs for the
'base paths', in order to remove the path config entirely.

Unfortunately, NamedBasePath and the Path object will take
additional work to untangle, because all of them are 'late bound'
- it is currently permitted to generate a NamedBasePath('checkout')
far, far before the checkout path is set (and many recipes do this).
This late binding is the core source of complexity here.

Change-Id: I59d4e9c50a8587ba4be97926398d5b48803acaa4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5097989
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Rob Mohr <mohrr@google.com>