[defer] Allow suppressing results in contexts
Allow suppressing all previous results in a defer context, but only if
there are additional failures. This is useful when we know a step has
failed, but we can provide a better failure message for it by parsing
the logs of the step in question.

Change-Id: I1b25ba0f57bde41fe4c4db7ccb458e8d22658aa5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5171421
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Rob Mohr <mohrr@google.com>