[config.Static] Add carve-out for Path objects.
Normally Static will call `hash` on any value passed in as a way to
check that the value passed in is immutable (mutable types are not
supposed to implement `__hash__`).

In the upcoming Path implementation, Paths will still be immutable,
but ones based on the 'checkout' global variable path (with
assign-once semantics) will hard error if they are evaluated before
the 'checkout' Path was assigned. This evaluation occurs during
__hash__ as well, meaning that modules (mostly in build.git) which
use a checkout-based Path as part of the configuration will raise an
exception (since all config stuff is evaluated before any Path is
assigned to the 'checkout' global Path).

To work around this, we add a special exemption from the hash
checkout just for the `Path` type. As the comment in the CL says,
the config types are all slated for removal (and, indeed, have been
untouched for many years now), so we don't expect this to become a
problematic workaround.

The change in freeze is done with the same rationale -
freeze is attempting to ensure that the objects it outputs are
immutable. It directly supports dicts, sequences and sets, but
all other objects it 'passes through' after calling 'hash' on
them to ensure they are immutable. Path objects expose no public
APIs which allow mutation, and so are ok to skip the hash check
here as well. `freeze` is frequently used on lists, dicts, etc.
which contain Path types.

R=bpastene, bryner, sshrimp

Bug: 329113288
Change-Id: Ifc21d8d75a9b43cdc071943a398817bc29322c52
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5378469
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>