[warn] Increase scope of warnings to be process-global.
Also refactors the warning logic so that ANY code at any point in the
recipe lifecycle (module-scoped code, code outside of the module system)
is able to emit a warning.

This also makes issuing a warning a first-class function in the
recipe_api python module, removing the 'warning' client and making
the warning recipe module a very basic wrapper.

Finally, this changes the way warning Causes are reported. Instead of
aggregating causes on a per-test-case level, they are accumulated
globally for the entire test runner process and then dumped in a single
Outcome message at the very end of the runner lifetime. To support
indicating which tests touched which errors, we retain a simple set
of warning NAMES that is attached to each test case result.

Change-Id: Ie8811609b18f741d6900a9c41c92fc0d096d61a7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3674618
Reviewed-by: Rob Mohr <mohrr@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>