[file] Fix bug in read_proto.
Due to a copy/paste error, read_proto took 'encoding_kwargs', which
makes very little sense. The typical use for this is to pass things
like 'ignore_unknown_fields' to the json_format decoder.

However, in the previous implementation, it took 'encoding_kwargs'
and passed this to the decoder AND an encoder (for the log report).

There's not a strong need to control the encoded output (especially
since this encoded output can vary from the selected encoder in
order to avoid dumping binary proto to a text log), so this fixes
the API to just have decoding_kwargs instead.

R=fancl, mohrr

Change-Id: I4d15a8c7030c8bf06a825bc12d1eaf785819f750
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5552112
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>