Prevent joining paths with None.
Currently if a Path is joined with None either using the joinpath method
or the / operator, the None gets ignored because all falsey values get
skipped over. This is the appropriate action for an empty string, but
None should be rejected in order to behave more similarly to the pathlib
Path type.

Bug: 329113288
Change-Id: Iad1e38ed422dcddfabf9490c2c2f917e69c625c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5641413
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>