Add a 'current_results_url' helper to milo's recipe_module
This method will return the milo URL to view the results in the
current invocation, eg:
https://luci-milo.appspot.com/ui/inv/build-8742673107691290081

Most of the time this won't be useful since the current build will
already be viewed in milo, which already has the "test results" tab
that shows that. But in some cases we'll be running the recipe locally
inside an rdb invocation (https://crbug.com/40890773). In that case,
it might be useful for the recipe to be able to generate and serve
up the URL at various points. This helper should be able to do that.

Bug: 41492686
Change-Id: I85a0317f1af9a539df4eaee2f0edaa75dca9f8c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5699048
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Weiwei Lin <weiweilin@google.com>