[file] Check that chmod mode is a str
Check that the mode argument to api.file.chmod() is a str and not an
int. If called with an int it produces surprising behavior. For
example, 'api.file.chmod("chmod path", "path", 0o777)' results in
'chmod path --mode 511' because 0o777 is 511 in decimal. Rather than
trying to interpret the caller's intent when a number is passed in,
just require the mode argument be a string.

Bug: b/372288022
Change-Id: Ibaba4be9e78d3b124fb266cc8c0a575d77f40902
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5916886
Auto-Submit: Rob Mohr <mohrr@google.com>
Commit-Queue: Rob Mohr <mohrr@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>