Fix find and replace fail
I got a little zealous with find and replace over in a71bdfef4b and
replaced the `verify` arg to `requests.get` with `cert`.

As it happens, `requests` HAS a `cert` arg, but it's for client certs
so this is not what we wanted.

This change fixes it to call `verify` again (though another
alternative might be to choose something else in case using a name
that `requests` uses for another thing is too confusing. WDYT?

Bug: 40250422
Change-Id: I9683267c6a9ac04afdb6948435c1ecf8521275a7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/5926329
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>