[lint] Better handle imports with dots
Previously, a line like this would be flagged:

  import urllib.parse

But this one would not:

  import urllib.parse as urlparse

This is because in the first case, the top-level module object is
'urllib', which is not permitted. In the second case, the top-level
module object is 'urllib.parse', which is permitted.

If a top-level module object only has one attribute, and that attribute
is another module, we can just evaluate that module instead of the
top-level one. The nested module's '__name__' is still fully-qualified.

Tested by adding those two lines to a recipe before and after applying
this change and running './recipes.py lint'.

Change-Id: I105480404b4df801c1fe35c27e81e8d7a1966343
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/6149871
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Rob Mohr <mohrr@google.com>